017 - adiciona inativação de paciente e todos seus registros

+++ LabMedical/src/app/components/registers/patient/pacient-form/pacient-form.component.ts
@@ -157,12 +157,15 @@ export class PacientFormComponent implements OnInit {
 
 	onDeleteRegistration() {
 		if (!this.hasRecords) {
-			this.pacientService.deletePatient(this.pacientId).subscribe(
-				() => {
-					this.toastr.success('Paciente deletado com sucesso')
-					this.navigateAway()
-				}
-			)
+			let confirmed = confirm("Deseja excluir permanentemente o paciente?")
+			if (confirmed) {
+				this.pacientService.deletePatient(this.pacientId).subscribe(
+					() => {
+						this.toastr.success('Paciente deletado com sucesso')
+						this.navigateAway()
+					}
+				)
+			}
 		} else {
 			this.toastr.error('Paciente possui registros. Não foi possível deletar o paciente.')
 		}
